+++ passwd.py
@@ -5,7 +5,7 @@ import random
 import json
 import uuid
 
-dictionary = [
+alpha_dict = [
   "a",
   "b",
   "c",
@@ -30,18 +30,10 @@ dictionary = [
   "w",
   "x",
   "y",
-  "z",
-  "0",
-  "1",
-  "2",
-  "3",
-  "4",
-  "5",
-  "6",
-  "7",
-  "8",
-  "9"
+  "z"
 ]
+
+num_dict = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9" ]
 
 options = {
   "words": int(os.environ.get('words') or 4),
@@ -55,14 +47,16 @@ Store the alredy used password letters count
 in order to check for max repeating letters
 '''
 used_letters = {}
+have_number = False
+have_upper = False
 
 # Generate a random boolean
-def is_upper():
+def random_bool():
   return bool(random.getrandbits(1))
 
 # Extract a random letter or number from dictionary
 def random_letter():
-  letter = random.choice(dictionary)
+  letter = random.choice(alpha_dict)
   if letter in used_letters:
     if used_letters[letter] is options["max_repeats"]:
       return random_letter()
@@ -70,18 +64,33 @@ def random_letter():
       used_letters[letter] += 1
   else:
     used_letters[letter] = 1
-  if is_upper() : letter = letter.upper()
   return letter
 
-# Extract a random password
+# Replace a random character from a word in a word list with the uppercase
+# version or with a random number
+def replace(word_list, number=False):
+	word_id = random.randint(0, 2)
+	rnd_word = word_list[word_id]
+	letter_id = random.randint(0, len(rnd_word)) - 1
+	if number:
+		replacement = random.choice(num_dict)
+	else:
+		replacement = rnd_word[letter_id].upper()
+	rnd_word = rnd_word[:letter_id] + replacement + rnd_word[letter_id:]
+	word_list[word_id] = rnd_word
+	return word_list
+
+# Create random password
 def generate_password():
-  password = ""
-  for i in range (0, options["words"]):
-    for j in range(0, options["word_length"]):
-      password += random_letter()
-    password += options["separator"]
-  # removes the last unwanted separator
-  return password[:-1]
+	word_list = []
+	for i in range (0, options["words"]):
+		new_word = ""
+		for j in range(0, options["word_length"]):
+			new_word += random_letter()
+		word_list.append(new_word)
+	word_list = replace(word_list)
+	word_list = replace(word_list, True)
+	return options["separator"].join(str(x) for x in word_list)
 
 '''
 Generates a new random password then creates the script filter
@@ -102,5 +111,4 @@ script_filter_result = json.dumps({
     }
   ]
 })
-
 sys.stdout.write(script_filter_result)